requirements.txt
pypdf==3.12.1
python-dateutil==2.8.2
typer==0.9.0
colorama==0.4.4
shellingham==1.4.0
pydantic==2.0.3
rich==13.6.0